+++ Makefile.am
@@ -424,6 +424,7 @@ libuv_la_CFLAGS += -D_UNIX03_THREADS \
 libuv_la_LDFLAGS += -qXPLINK
 libuv_la_SOURCES += src/unix/pthread-fixes.c \
                     src/unix/pthread-barrier.c \
+                    src/unix/no-fsevents.c \
                     src/unix/os390.c \
                     src/unix/os390-syscalls.c \
                     src/unix/proctitle.c
